requirements.txt
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
djangorestframework==3.15.2